Use half-close when closing the socking in HttpServerEarlyClientCloseTest.
This should fix the flankyness of the test under Windows. The lesson to
learn from this, is that a Socket.close() is no guaranteed to have sent
all data to the remote peer, even though
Socket.outputStream.onNoPendingWrites have been issued, on Windows.

We might have to visit this later on, if it becomes a pattern other
users fall into.

BUG=http://code.google.com/p/dart/issues/detail?id=2551
TEST=

Review URL: https://chromiumcodereview.appspot.com//10187002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@6868 260f80e4-7a28-3924-810f-c04153c831b5