Fix for unsafe cast in text search.
Caught by Terry here: http://code.google.com/p/dart/issues/detail?id=2760

The net-net is that the added behavior to trigger a text search on a carriage return isn't safe when the header element has a pointer to a PreviousPicksProvider.  This fixes that.
Review URL: https://chromiumcodereview.appspot.com//10377073

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@7515 260f80e4-7a28-3924-810f-c04153c831b5